requirements._all.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
efficientnet-pytorch==0.7.1
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.5.0
pretrainedmodels==0.7.4
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
segmentation-models-pytorch==0.3.2
six==1.16.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
timm==0.6.12
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
